+++ crates/agnetd/src/cmd_inference.rs
@@ -98,10 +98,11 @@ fn request_inference(
 
     let providers = load_all_providers(state.db());
     let pricing_info = providers.iter().find_map(|p| p.find_model(model)).map(|m| {
-        let cost = (estimated_tokens as u64 * m.input_price_per_million.0 / 1_000_000).max(1);
+        let cost =
+            ((estimated_tokens as u128 * m.input_price_per_million.0 / 1_000_000).max(1)) as u64;
         serde_json::json!({
-            "input_price_per_mtok": m.input_price_per_million.0,
-            "output_price_per_mtok": m.output_price_per_million.0,
+            "input_price_per_mtok": m.input_price_per_million.0 as u64,
+            "output_price_per_mtok": m.output_price_per_million.0 as u64,
             "estimated_cost": cost,
         })
     });
@@ -275,8 +276,10 @@ fn show_pricing(
             capabilities: vec!["chat".to_string()],
         });
 
-    let input_cost = (input_tokens as u64) * model_info.input_price_per_million.0 / 1_000_000;
-    let output_cost = (output_tokens as u64) * model_info.output_price_per_million.0 / 1_000_000;
+    let input_cost =
+        ((input_tokens as u128) * model_info.input_price_per_million.0 / 1_000_000) as u64;
+    let output_cost =
+        ((output_tokens as u128) * model_info.output_price_per_million.0 / 1_000_000) as u64;
     let total = input_cost.saturating_add(output_cost);
     let total_cost =
         if total == 0 && (input_tokens > 0 || output_tokens > 0) { 1u64 } else { total };
